third_party/blink/web_tests/virtual/scoped-custom-element-registry/external/wpt/custom-elements/registries/ShadowRoot-innerHTML-expected.txt

This is a testharness.js-based test.
[FAIL] innerHTML on a connected shadow root should use the associated scoped registry
  assert_true: expected true got false
[FAIL] innerHTML on a connected shadow root should not upgrade a custom element inside a template element
  assert_equals: expected "SomeElement" but got "SomeElement1"
Harness: the test ran to completion.
